+++ src/extension.ts
@@ -17,10 +17,14 @@ export function activate(context: vscode.ExtensionContext) {
 
 	// Register commands
 	let command = vscode.commands.registerCommand('base64viewer.decodeBase64', () => {
-		vscode.window.showInputBox({ prompt: messages.general.prompt.decode }).then(
-			(base64String) => decodeAndDisplay(extensionRoot, xss(base64String)),
-			(reason) => showErrorPopup(reason),
-		);
+		vscode.window
+			.showInputBox({
+				prompt: messages.general.prompt.decode,
+			})
+			.then(
+				(base64String) => decodeAndDisplay(extensionRoot, xss(base64String)),
+				(reason) => showErrorPopup(reason),
+			);
 	});
 	context.subscriptions.push(command);